+++ modules/@apostrophecms/piece-type/ui/apos/components/AposDocsManager.vue
@@ -137,6 +137,7 @@ export default {
     return {
       modal: {
         active: false,
+        triggerFocusRefresh: 0,
         type: 'overlay',
         showModal: false
       },
@@ -230,6 +231,7 @@ export default {
     this.modal.active = true;
     await this.getPieces();
     await this.getAllPiecesTotal();
+    this.modal.triggerFocusRefresh++;
 
     apos.bus.$on('content-changed', this.getPieces);
     apos.bus.$on('command-menu-manager-create-new', this.create);