+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 Pages/Docs Manager and Doc Editor modal now have better keyboard accessibility.
 They keep the focus on elements inside modals and give it back to their parent modal when closed.
 
+### Fixes
+- Fixes a problem in the rich text editor where the slash would not be deleted after item selectin from the insert menu.
+
 ## 3.52.0 (2023-07-06)
 
 ### Changes