+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Accessibility improved for navigation inside modals and various UI elements.  
 Pages/Docs Manager and Doc Editor modal now have better keyboard accessibility.
 They keep the focus on elements inside modals and give it back to their parent modal when closed.
+* Adds support for a new `if` property in `addContextOperation` in order to show or not a context operation based on the current document properties. 
 
 ### Fixes
 - Fixes a problem in the rich text editor where the slash would not be deleted after item selectin from the insert menu.