+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## UNRELEASED
 
+### Adds
+
+* Accessibility improved for navigation inside modals and various UI elements.  
+Pages/Docs Manager and Doc Editor modal now have better keyboard accessibility.
+They keep the focus on elements inside modals and give it back to their parent modal when closed.
+* Adds support for a new `if` property in `addContextOperation` in order to show or not a context operation based on the current document properties. 
+
+### Fixes
+- Fixes a problem in the rich text editor where the slash would not be deleted after item selectin from the insert menu.
+
+## 3.52.0 (2023-07-06)
+
 ### Changes
 
 * Foreign widget UI no longer uses inverted theme styles.
@@ -22,6 +34,7 @@ by an `if` condition fails to satisfy a condition such as `min` or `max`
 or is otherwise invalid. Instead the invalid value is discarded for safety.
 Note that `required` has always been ignored when an `if` condition is not
 satisfied.
+* Errors thrown in `@apostrophecms/login:afterSessionLogin` event handlers are now properly passed back to Passport as such, avoiding a process restart.
 
 ## 3.51.1 (2023-06-23)